Apply sample normalisation to loud samples
Adds -4db headroom on all samples which exceeded this amount. This
roughly matches what stable uses for hitsounds, so I figured it's a good
starting point.

I have only changed wav files, as most of the mp3 samples are not
responsible for concurrent playback and therefore less of an issue.