Fix localisation resx file casing
The .NET build toolchain expects that the dialect part of the language
tag in the filename of a resx file should be upper-cased. However, some
resx files with the dialect part were committed as fully lower-cased,
which causes compilation failures on filesystems/OSes on which filenames
are case-sensitive, which is primarily Linux.

Move the offending files so that they use the proper upper-cased suffix.